+++ src/main/java/io/xdag/core/Blockchain.java
@@ -24,12 +24,13 @@
 
 package io.xdag.core;
 
-import io.xdag.listener.Listener;
 import java.util.List;
 import java.util.Map;
-import org.apache.tuweni.bytes.Bytes;
+
 import org.apache.tuweni.bytes.Bytes32;
 import org.hyperledger.besu.crypto.KeyPair;
+
+import io.xdag.listener.Listener;
 
 public interface Blockchain {
 
@@ -51,8 +52,6 @@ public interface Blockchain {
     List<Block> listMainBlocks(int count);
 
     List<Block> listMinedBlocks(int count);
-
-    Map<Bytes, Integer> getMemOurBlocks();
 
     XdagStats getXdagStats();
