+++ src/main/java/io/xdag/crypto/Base58.java
@@ -81,7 +81,7 @@ public class Base58 {
 
 
     public static byte[] decode(String input) throws AddressFormatException {
-        if (input.length() == 0) {
+        if (input.isEmpty()) {
             return new byte[0];
         }
         // Convert the base58-encoded ASCII chars to a base58 byte sequence (base58 digits).
