+++ src/main/java/io/xdag/crypto/MnemonicUtils.java
@@ -60,7 +60,7 @@ public class MnemonicUtils {
      * The mnemonic must encode entropy in a multiple of 32 bits. With more entropy security is
      * improved but the sentence length increases. We refer to the initial entropy length as ENT.
      * The allowed size of ENT is 128-256 bits.
-     *
+     * <p>
      * <h3>Mnemonic generation algorithm</h3>
      *
      * Given a randomly generated initial entropy of size ENT, first a checksum is generated by
