+++ src/main/java/io/xdag/net/message/consensus/SumReplyMessage.java
@@ -23,17 +23,12 @@
  */
 package io.xdag.net.message.consensus;
 
-import java.math.BigInteger;
-
-import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.bytes.MutableBytes;
 
-import io.xdag.core.SimpleEncoder;
+import io.xdag.utils.SimpleEncoder;
 import io.xdag.core.XdagStats;
 import io.xdag.net.NetDB;
 import io.xdag.net.message.MessageCode;
-import io.xdag.utils.BytesUtils;
-import io.xdag.utils.Numeric;
 import io.xdag.utils.SimpleDecoder;
 import lombok.Getter;
 import lombok.Setter;
