+++ src/main/java/io/xdag/net/message/p2p/PongMessage.java
@@ -23,7 +23,7 @@
  */
 package io.xdag.net.message.p2p;
 
-import io.xdag.core.SimpleEncoder;
+import io.xdag.utils.SimpleEncoder;
 import io.xdag.net.message.Message;
 import io.xdag.net.message.MessageCode;
 import io.xdag.utils.SimpleDecoder;
