+++ src/test/java/io/xdag/net/message/p2p/HelloMessageTest.java
@@ -28,13 +28,13 @@ import static junit.framework.TestCase.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import org.hyperledger.besu.crypto.KeyPair;
+import org.hyperledger.besu.crypto.SecureRandomProvider;
 import org.junit.Test;
 
 import io.xdag.config.Config;
 import io.xdag.config.DevnetConfig;
 import io.xdag.crypto.Keys;
 import io.xdag.crypto.SampleKeys;
-import io.xdag.crypto.SecureRandomUtils;
 import io.xdag.crypto.Sign;
 import io.xdag.net.CapabilityTreeSet;
 import io.xdag.net.Peer;
@@ -45,11 +45,11 @@ public class HelloMessageTest {
     public void testCodec() {
         Config config = new DevnetConfig();
 
-        KeyPair key = KeyPair.create(SampleKeys.SRIVATE_KEY, Sign.CURVE, Sign.CURVE_NAME);;
+        KeyPair key = KeyPair.create(SampleKeys.SRIVATE_KEY, Sign.CURVE, Sign.CURVE_NAME);
         String peerId = toBase58(Keys.toBytesAddress(key));
         HelloMessage msg = new HelloMessage(config.getNodeSpec().getNetwork(), config.getNodeSpec().getNetworkVersion(), peerId, 8001,
                 config.getClientId(), config.getClientCapabilities().toArray(), 2,
-                SecureRandomUtils.secureRandom().generateSeed(InitMessage.SECRET_LENGTH), key);
+                SecureRandomProvider.publicSecureRandom().generateSeed(InitMessage.SECRET_LENGTH), key);
         assertTrue(msg.validate(config));
 
         msg = new HelloMessage(msg.getBody());
