+++ src/test/java/io/xdag/utils/BytesTest.java
@@ -25,13 +25,14 @@ package io.xdag.utils;
 
 import static org.junit.Assert.assertEquals;
 
-import io.xdag.crypto.Hash;
 import java.nio.charset.StandardCharsets;
+
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.bytes.MutableBytes32;
-import org.junit.Assert;
 import org.junit.Test;
+
+import io.xdag.crypto.Hash;
 
 public class BytesTest {
